--- src/components/signup/index.tsx
@@ -1,9 +1,0 @@
-function Signup() {
-  return (
-    <form action="">
-      <input type="text" />
-    </form>
-  );
-}
-
-export default Signup;
