+++ src/pages/_app.tsx
@@ -1,15 +1,27 @@
 import '@/styles/globals.css';
+import { NextPage } from 'next';
 import type { AppProps } from 'next/app';
 import { Noto_Sans_KR } from 'next/font/google';
+import { ReactElement, ReactNode } from 'react';
 
 const noto = Noto_Sans_KR({
   subsets: ['latin'],
 });
 
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <main className={noto.className}>
+export type NextPageWithLayout<P = unknown, IP = P> = NextPage<P, IP> & {
+  getLayout?: (page: ReactElement) => ReactNode;
+};
+
+type AppPropsWithLayout = AppProps & {
+  Component: NextPageWithLayout;
+};
+
+export default function App({ Component, pageProps }: AppPropsWithLayout) {
+  const getLayout = Component.getLayout ?? ((page) => page);
+
+  return getLayout(
+    <div className={noto.className}>
       <Component {...pageProps} />
-    </main>
+    </div>,
   );
 }
